requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
alabaster==0.7.12
APScheduler==3.9.1.post1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
bidict==0.22.1
black==23.3.0
bleach==5.0.1
blinker==1.6.2
boltons==21.0.0
boto3==1.28.3
botocore==1.31.3
CacheControl==0.12.14
cachetools==5.3.1
certifi==2022.5.18.1
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
chatgpt==2.2212.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.5
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==2.2.0
constantly==15.1.0
cors==1.0.1
cryptography==36.0.2
cssselect==1.1.0
cycler==0.11.0
dask==2022.10.2
dataclasses-json==0.5.7
datasets==2.12.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.0.7
dill==0.3.6
distlib==0.3.6
dnspython==2.3.0
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.0.0
fastjsonschema==2.16.1
fbmessenger==6.0.0
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
fonttools==4.29.1
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
gast==0.4.0
gevent==23.7.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
-e git+https://github.com/AntonOsika/gpt-engineer.git@d91384c2aa21d3c617d4f7a387c3ad7b4b0823c6#egg=gpt_engineer
greenlet==2.0.2
grpcio==1.56.0
gunicorn==21.2.0
h5py==3.9.0
httptools==0.6.0
huggingface-hub==0.15.1
humanfriendly==10.0
hyperlink==21.0.0
identify==2.5.24
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
iniconfig==2.0.0
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
jax==0.4.13
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
jsonpickle==3.0.1
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.12.0
kiwisolver==1.3.2
libclang==16.0.0
lkml==1.1.0
locket==1.0.0
lxml==4.9.0
magic-filter==1.0.9
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mattermostwrapper==2.2
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.2.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==5.2.0
multiprocess==0.70.14
mypy==1.3.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.6.3
nltk==3.8.1
nodeenv==1.8.0
notebook==6.4.12
numpy==1.22.2
oauthlib==3.2.2
openai==0.27.8
openpyxl==3.0.10
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==20.9
pamqp==3.2.1
pandas==1.4.1
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.3
partd==1.4.0
pathspec==0.11.1
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.8.1
pluggy==1.2.0
portalocker==2.7.0
pre-commit==3.3.3
prometheus-client==0.14.1
prompt-toolkit==3.0.28
Protego==0.2.1
protobuf==4.23.3
psutil==5.9.2
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.10.9
PyDispatcher==2.0.5
pydot==1.4.2
Pygments==2.13.0
PyJWT==2.4.0
pykwalify==1.8.0
pymongo==4.3.3
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyreadline3==3.4.1
pyreadstat==1.1.4
pyrsistent==0.18.1
PySocks==1.7.1
pyTelegramBotAPI==4.5.1
pytest==7.3.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2021.3
pywin32==304
pywinpty==2.0.7
PyYAML==5.4.1
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
questionary==1.10.0
queuelib==1.6.2
randomname==0.1.5
rasa==3.6.2
rasa-sdk==3.6.1
redis==4.6.0
regex==2022.10.31
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rich==13.4.2
rocketchat-API==1.30.0
rsa==4.9
ruamel.base==1.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
rudder-sdk-python==2.0.2
ruff==0.0.272
s3transfer==0.6.1
safetensors==0.3.1
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.0.2
scipy==1.8.0
Scrapy==2.6.1
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.8.0
sentry-sdk==1.14.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0
sklearn-crfsuite==0.3.6
slack-sdk==3.21.3
smmap==5.0.0
snowballstemmer==2.2.0
snowflake==0.0.3
snowflake-connector-python==2.7.9
snowflake-sqlalchemy==1.3.4
soupsieve==2.3.2.post1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.20
stack-data==0.5.0
statsmodels==0.13.2
structlog==23.1.0
structlog-sentry==2.0.3
sympy==1.12
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.15.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.1.1
tldextract==3.3.0
tls-client==0.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.4.0
transformers==4.30.2
twilio==8.2.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typer==0.9.0
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.2.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==1.26.16
vaderSentiment==3.3.2
virtualenv==20.23.1
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.3.6
widgetsnbextension==4.0.3
wrapt==1.14.1
xgboost==1.6.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.8.1
zope.event==5.0
zope.interface==5.4.0
